+++ lib/truffle.rb
@@ -26,6 +26,7 @@ require_relative "truffle/providers/anthropic_stream"
 require_relative "truffle/providers/google"
 require_relative "truffle/providers/google_stream"
 require_relative "truffle/agent"
+require_relative "truffle/agent/tool_execution"
 require_relative "truffle/compaction"
 require_relative "truffle/tools"